Extract QTI import payload construction into a pure function
